Layout: Movie info page upgraded
