+++ src/routes/Routes.jsx
@@ -32,8 +32,8 @@ import PopularTvs from '../pages/home/popularTvs/PopularTvs';
 import PopularTv from '../pages/home/popularTvs/popularTv';
 import LiveTv from '../pages/liveTv/LiveTv';
 import Login from '../pages/login/Login';
-import MovieInfo from '../pages/movieInfo/MovieInfo';
 import Movies from '../pages/movies/Movies';
+import MovieInfo from '../pages/movies/movieInfo/MovieInfo';
 import Payment from '../pages/payment/Payment';
 import Podcast from '../pages/podcast/Podcast';
 import PrivacyPolicy from '../pages/policy/PrivacyPolicy';
